shuffle organization of adapters and protocols
