+++ jobbers/adapters/protocols.py
@@ -1,0 +1,143 @@
+"""
+Protocol definitions for all pluggable adapters.
+
+Routing:
+- `RoutingBackendReadOnlyError` — raised by read-only backends on write operations.
+- `RoutingBackendProtocol` — interface all routing backends must implement.
+
+Task storage / dead-letter queue:
+- `TaskAdapterProtocol` — interface all task adapters must implement.
+- `DeadQueueProtocol` — interface all dead-letter queue implementations must implement.
+"""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, runtime_checkable
+
+from typing_extensions import Protocol
+
+if TYPE_CHECKING:
+    import datetime as dt
+    from collections.abc import AsyncGenerator
+    from typing import Any
+
+    from redis.asyncio.client import Pipeline
+    from ulid import ULID
+
+    from jobbers.models.dag import DAGRunPagination
+    from jobbers.models.queue_config import QueueConfig
+    from jobbers.models.task import Task, TaskPagination
+    from jobbers.models.task_routing import RoutingConfig
+
+
+class RoutingBackendReadOnlyError(Exception):
+    """Raised when a write operation is attempted on a read-only routing backend."""
+
+
+@runtime_checkable
+class RoutingBackendProtocol(Protocol):
+    """Interface all routing backends must implement."""
+
+    # Queue CRUD
+    async def get_queue_config(self, queue: str) -> QueueConfig | None: ...
+    async def save_queue_config(self, queue_config: QueueConfig) -> None: ...
+    async def delete_queue(self, queue_name: str) -> None: ...
+    async def get_all_queues(self) -> list[str]: ...
+
+    # Role CRUD
+    async def get_queues(self, role: str) -> set[str]: ...
+    async def save_role(self, role: str, queues_set: set[str]) -> str: ...
+    async def get_all_roles(self) -> list[str]: ...
+    async def delete_role(self, role: str) -> None: ...
+
+    # Change detection (ULID-stamped refresh tags drive pub/sub worker refresh)
+    async def get_refresh_tag(self, role: str) -> ULID: ...
+    async def bump_refresh_tag(self, role: str) -> str: ...
+    async def bump_refresh_tags_for_queue(self, queue_name: str) -> list[str]: ...
+
+    # Task routing config CRUD
+    async def get_routing_config(self, task_name: str, task_version: int) -> RoutingConfig | None: ...
+    async def save_routing_config(self, routing_config: RoutingConfig) -> None: ...
+    async def delete_routing_config(self, task_name: str, task_version: int) -> bool: ...
+
+
+@runtime_checkable
+class TaskAdapterProtocol(Protocol):  # pragma: no cover
+    """Interface for task storage and querying."""
+
+    # -- key helpers (both implementations use the same Redis key names) ----
+    TASKS_BY_QUEUE: Any
+    TASK_DETAILS: Any
+    HEARTBEAT_SCORES: Any
+    TASK_BY_TYPE_IDX: Any
+    QUEUE_RATE_LIMITER: Any
+    DLQ_MISSING_DATA: str
+    DAG_RUNS: str
+    DAG_RUN_TASKS: Any
+
+    # -- write path ----------------------------------------------------------
+    async def submit_task(self, task: Task) -> bool: ...
+    async def submit_rate_limited_task(self, task: Task, queue_config: QueueConfig) -> bool: ...
+    def stage_save(self, pipe: Pipeline, task: Task) -> None: ...
+    def stage_requeue(self, pipe: Pipeline, task: Task) -> None: ...
+    def stage_submit_task(self, pipe: Pipeline, task: Task) -> None: ...
+
+    # -- read path -----------------------------------------------------------
+    async def get_task(self, task_id: ULID) -> Task | None: ...
+    async def get_tasks_bulk(self, task_ids: list[ULID]) -> list[Task | None]: ...
+    async def read_for_watch(self, pipe: Pipeline, task_id: ULID) -> Task | None: ...
+    async def task_exists(self, task_id: ULID) -> bool: ...
+    async def get_active_tasks(self, queues: set[str]) -> list[Task]: ...
+    def get_stale_tasks(self, queues: set[str], stale_time: dt.timedelta) -> AsyncGenerator[Task, None]: ...
+    async def get_all_tasks(self, pagination: TaskPagination) -> list[Task]: ...
+    async def get_next_task(self, queues: set[str], pop_timeout: int) -> Task | None: ...
+
+    # -- queue management ----------------------------------------------------
+    def stage_remove_from_queue(self, pipe: Pipeline, task: Task) -> None: ...
+    async def update_task_heartbeat(self, task: Task) -> None: ...
+    async def remove_task_heartbeat(self, task: Task) -> None: ...
+
+    # -- dag fan-in ----------------------------------------------------------
+    def stage_init_fan_in(
+        self, pipe: Pipeline, fan_in_key: str, predecessor_ids: set[ULID], ttl: int = 86400
+    ) -> None: ...
+    async def init_fan_in(self, fan_in_key: str, predecessor_ids: set[ULID], ttl: int = 86400) -> None: ...
+    async def fan_in_complete(self, fan_in_key: str, task_id: ULID) -> int: ...
+    async def get_fan_in_members(self, fan_in_key: str) -> list[ULID]: ...
+
+    # -- dag run index -------------------------------------------------------
+    async def get_dag_runs(
+        self, pagination: DAGRunPagination
+    ) -> tuple[list[tuple[ULID, dt.datetime]], int]: ...
+    async def get_dag_run(self, dag_run_id: ULID) -> tuple[dt.datetime, list[ULID]] | None: ...
+    async def clean_dag_runs(self, now: dt.datetime, max_age: dt.timedelta) -> None: ...
+
+    # -- lifecycle -----------------------------------------------------------
+    async def ensure_index(self) -> None: ...
+    async def clean_terminal_tasks(self, now: dt.datetime, max_age: dt.timedelta) -> None: ...
+    async def clean(
+        self,
+        queues: set[bytes],
+        now: dt.datetime,
+        min_queue_age: dt.datetime | None,
+        max_queue_age: dt.datetime | None,
+    ) -> None: ...
+
+
+class DeadQueueProtocol(Protocol):  # pragma: no cover
+    """Interface for dead letter queue operations."""
+
+    async def ensure_index(self) -> None: ...
+    def stage_add(self, pipe: Pipeline, task: Task, failed_at: dt.datetime) -> None: ...
+    def stage_remove(self, pipe: Pipeline, task_id: ULID, queue: str, name: str) -> None: ...
+    async def get_history(self, task_id: str) -> list[dict[str, Any]]: ...
+    async def get_by_ids(self, task_ids: list[str]) -> list[Task]: ...
+    async def get_by_filter(
+        self,
+        queue: str | None,
+        task_name: str | None,
+        task_version: int | None,
+        limit: int,
+    ) -> list[Task]: ...
+    async def remove_many(self, task_ids: list[str]) -> None: ...
+    async def clean(self, earlier_than: dt.datetime) -> None: ...
