+++ jobbers/runners/worker_proc.py
@@ -94,7 +94,7 @@ def run() -> None:
 
     if args.static_config:
         from jobbers import db
-        from jobbers.adapters.static_routing import StaticRoutingBackend
+        from jobbers.adapters.static import StaticRoutingBackend
 
         db.register_routing_backend(StaticRoutingBackend.from_file(args.static_config))
